Switch to dot.access syntax in Obstacle also
It's making me very twitchy if I want to extract a field to another module and it breaks everything. But pretty sure find/replace works fine when that happens

For now let's lean twords dot.syntax in our own module, and public getters for outsiders. Whatever, this is fine for now stop obsessing