Inline the getters/setters/updaters
In the spirit of not reaching for them defensivly, just inline them here

Maybe the function names are more readable at a glance (and I AM doing that with lockedIn? and legalMove?), but it's fine here. Look at the anon function being called, oh this is the Move line. Oh, this is the Lockin line

Good to go