The Game Lobby genserver will start up when our app does
Very cool. We only need 1 lobby, and it's "state"'y. Doesn't really need to be db, just a list of all running games and if the entire server ever crashed we'd just lose that list of games

Later games and which ones are active can go into a db maybe, for now this is perfect use of the "medium-duration" state genserver sort of thing